+++ src/rmq-node.lisp
@@ -169,8 +169,13 @@ return :success t with the :result if things go well"
       (iter:iterate
         (iter:for item in result)
         (handler-case
-            (progn (send-message node (rmq-node/fail-queue node) (rmq-message-body item))
-                   (nack-message node item nil))
+            ;; NOTE: not all failures result in a 'valid' message, that is a message
+            ;; that came from rmq.
+            ;; In this case we can just drop it on the floor.
+            ;; TODO: test this code path
+            (when (rmq-message-delivery-tag item)
+              (send-message node (rmq-node/fail-queue node) (rmq-message-body item))
+              (nack-message node item nil))
           (rabbitmq-library-error (c)
             (v:error :node.event-loop "rmq-error- failed to place item (~d): ~a"
                        (rabbitmq-library-error/error-code c)